backend/requirements.txt
fastapi
uvicorn
requests
PyMuPDF
openpyxl
python-multipart
Pillow
pandas
python-dotenv
openai
